+++ docs/changelog.md
@@ -3,6 +3,13 @@
 * Upgraded Ubuntu machine image in Circle CI configuration to version `ubuntu-2004:202201-02`
 * Upgraded `ansi-regex` to v5.0.1 due the vulnerability found by dependabot in v5.0.0
 
+# 0.5.19
+### Features
+* Added wait-for-availability checks to the indexer, wallet, contract-helper service, explorer-wamp, and explorer
+
+### Fixes
+* Fixed an issue with the indexer database not getting created
+
 # 0.5.18
 ### Features
 * Implemented static public ports